+++ public_transport_watcher/utils/__init__.py
@@ -6,6 +6,7 @@ from .get_env_variable import get_env_variable
 from .get_query_result import get_query_result
 from .get_sql_query import get_sql_query
 from .get_trafic_api import fetch_generic_api_data
+from .get_cache_utils import is_cache_valid, load_from_cache, save_to_cache, get_cache_info
 
 __all__ = [
     "fetch_generic_api_data",
@@ -16,4 +17,8 @@ __all__ = [
     "get_env_variable",
     "get_sql_query",
     "get_query_result",
+    "is_cache_valid",
+    "load_from_cache",
+    "save_to_cache",
+    "get_cache_info",
 ]